优化：将字典里的NSNumben以字符串类型取出

+++ RichCategory/Classes/NSDictionary+SafeObject.h
@@ -12,7 +12,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface NSDictionary (SafeObject)
 
-- (int)safeIntForKey:(NSString *)key;
+- (NSInteger)safeIntForKey:(NSString *)key;
 
 - (NSString *)safeStringForKey:(NSString *)key;
